@@ -18,10 +18,10 @@ public class ErrorResourceSerializer extends JsonSerializer<ErrorResource> {
     gen.writeStartObject();
     gen.writeObjectFieldStart("errors");
     for (FieldErrorResource fieldErrorResource : value.getFieldErrors()) {
-      if (!json.containsKey(fieldErrorResource.getField())) {
-        json.put(fieldErrorResource.getField(), new ArrayList<String>());
+      if (!json.containsKey(fieldErrorResource.field())) {
+        json.put(fieldErrorResource.field(), new ArrayList<String>());
       }
-      json.get(fieldErrorResource.getField()).add(fieldErrorResource.getMessage());
+      json.get(fieldErrorResource.field()).add(fieldErrorResource.message());
     }
     for (Map.Entry<String, List<String>> pair : json.entrySet()) {
       gen.writeArrayFieldStart(pair.getKey());
