@@ -1,15 +1,6 @@
 package io.spring.api.exception;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-@Getter
-@AllArgsConstructor
-public class FieldErrorResource {
-  private String resource;
-  private String field;
-  private String code;
-  private String message;
-}
+public record FieldErrorResource(String resource, String field, String code, String message) {}
