@@ -9,6 +9,7 @@ import io.spring.core.user.User;
 import io.spring.infrastructure.mybatis.readservice.ArticleFavoritesReadService;
 import io.spring.infrastructure.mybatis.readservice.ArticleReadService;
 import io.spring.infrastructure.mybatis.readservice.UserRelationshipQueryService;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -17,7 +18,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import lombok.AllArgsConstructor;
-import org.joda.time.DateTime;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -55,7 +55,7 @@ public class ArticleQueryService {
       String tag,
       String author,
       String favoritedBy,
-      CursorPageParameter<DateTime> page,
+      CursorPageParameter<Instant> page,
       User currentUser) {
     List<String> articleIds =
         articleReadService.findArticlesWithCursor(tag, author, favoritedBy, page);
@@ -78,7 +78,7 @@ public class ArticleQueryService {
   }
 
   public CursorPager<ArticleData> findUserFeedWithCursor(
-      User user, CursorPageParameter<DateTime> page) {
+      User user, CursorPageParameter<Instant> page) {
     List<String> followdUsers = userRelationshipQueryService.followedUsers(user.getId());
     if (followdUsers.size() == 0) {
       return new CursorPager<>(new ArrayList<>(), page.getDirection(), false);
@@ -140,7 +140,7 @@ public class ArticleQueryService {
     articles.forEach(
         articleData -> {
           if (followingAuthors.contains(articleData.getProfileData().getId())) {
-            articleData.getProfileData().setFollowing(true);
+            articleData.getProfileData().updateFollowing(true);
           }
         });
   }
@@ -152,10 +152,10 @@ public class ArticleQueryService {
     Map<String, Integer> countMap = new HashMap<>();
     favoritesCounts.forEach(
         item -> {
-          countMap.put(item.getId(), item.getCount());
+          countMap.put(item.id(), item.count());
         });
     articles.forEach(
-        articleData -> articleData.setFavoritesCount(countMap.get(articleData.getId())));
+        articleData -> articleData.updateFavoritesCount(countMap.get(articleData.getId())));
   }
 
   private void setIsFavorite(List<ArticleData> articles, User currentUser) {
@@ -167,17 +167,17 @@ public class ArticleQueryService {
     articles.forEach(
         articleData -> {
           if (favoritedArticles.contains(articleData.getId())) {
-            articleData.setFavorited(true);
+            articleData.updateFavorited(true);
           }
         });
   }
 
   private void fillExtraInfo(String id, User user, ArticleData articleData) {
-    articleData.setFavorited(articleFavoritesReadService.isUserFavorite(user.getId(), id));
-    articleData.setFavoritesCount(articleFavoritesReadService.articleFavoriteCount(id));
+    articleData.updateFavorited(articleFavoritesReadService.isUserFavorite(user.getId(), id));
+    articleData.updateFavoritesCount(articleFavoritesReadService.articleFavoriteCount(id));
     articleData
         .getProfileData()
-        .setFollowing(
+        .updateFollowing(
             userRelationshipQueryService.isUserFollowing(
                 user.getId(), articleData.getProfileData().getId()));
   }
