@@ -3,26 +3,15 @@ package io.spring.application.article;
 import com.fasterxml.jackson.annotation.JsonRootName;
 import java.util.List;
 import javax.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
 @JsonRootName("article")
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class NewArticleParam {
-  @NotBlank(message = "can't be empty")
-  @DuplicatedArticleConstraint
-  private String title;
-
-  @NotBlank(message = "can't be empty")
-  private String description;
-
-  @NotBlank(message = "can't be empty")
-  private String body;
-
-  private List<String> tagList;
+public record NewArticleParam(
+    @NotBlank(message = "can't be empty") @DuplicatedArticleConstraint String title,
+    @NotBlank(message = "can't be empty") String description,
+    @NotBlank(message = "can't be empty") String body,
+    List<String> tagList) {
+  public static NewArticleParam of(
+      String title, String description, String body, List<String> tagList) {
+    return new NewArticleParam(title, description, body, tagList);
+  }
 }
