@@ -1,16 +1,6 @@
 package io.spring.application.article;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
 @JsonRootName("article")
-public class UpdateArticleParam {
-  private String title = "";
-  private String body = "";
-  private String description = "";
-}
+public record UpdateArticleParam(String title, String body, String description) {}
