@@ -2,13 +2,11 @@ package io.spring.application.data;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.spring.application.DateTimeCursor;
+import java.time.Instant;
 import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.joda.time.DateTime;
+import lombok.*;
 
-@Data
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class ArticleData implements io.spring.application.Node {
@@ -19,8 +17,8 @@ public class ArticleData implements io.spring.application.Node {
   private String body;
   private boolean favorited;
   private int favoritesCount;
-  private DateTime createdAt;
-  private DateTime updatedAt;
+  private Instant createdAt;
+  private Instant updatedAt;
   private List<String> tagList;
 
   @JsonProperty("author")
@@ -29,5 +27,13 @@ public class ArticleData implements io.spring.application.Node {
   @Override
   public DateTimeCursor getCursor() {
     return new DateTimeCursor(updatedAt);
+  }
+
+  public void updateFavorited(boolean favorited) {
+    this.favorited = favorited;
+  }
+
+  public void updateFavoritesCount(int favoritesCount) {
+    this.favoritesCount = favoritesCount;
   }
 }
