@@ -1,9 +1,3 @@
 package io.spring.application.data;
 
-import lombok.Value;
-
-@Value
-public class ArticleFavoriteCount {
-  private String id;
-  private Integer count;
-}
+public record ArticleFavoriteCount(String id, Integer count) {}
