@@ -3,24 +3,12 @@ package io.spring.application.user;
 import com.fasterxml.jackson.annotation.JsonRootName;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
 @JsonRootName("user")
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterParam {
-  @NotBlank(message = "can't be empty")
-  @Email(message = "should be an email")
-  @DuplicatedEmailConstraint
-  private String email;
-
-  @NotBlank(message = "can't be empty")
-  @DuplicatedUsernameConstraint
-  private String username;
-
-  @NotBlank(message = "can't be empty")
-  private String password;
-}
+public record RegisterParam(
+    @NotBlank(message = "can't be empty")
+        @Email(message = "should be an email")
+        @DuplicatedEmailConstraint
+        String email,
+    @NotBlank(message = "can't be empty") @DuplicatedUsernameConstraint String username,
+    @NotBlank(message = "can't be empty") String password) {}
