@@ -1,10 +1,10 @@
 package io.spring.core.comment;
 
+import java.time.Instant;
 import java.util.UUID;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.joda.time.DateTime;
 
 @Getter
 @NoArgsConstructor
@@ -14,13 +14,13 @@ public class Comment {
   private String body;
   private String userId;
   private String articleId;
-  private DateTime createdAt;
+  private Instant createdAt;
 
   public Comment(String body, String userId, String articleId) {
     this.id = UUID.randomUUID().toString();
     this.body = body;
     this.userId = userId;
     this.articleId = articleId;
-    this.createdAt = new DateTime();
+    this.createdAt = Instant.now();
   }
 }
