@@ -36,12 +36,12 @@ public class MeDatafetcher {
     io.spring.core.user.User user = (io.spring.core.user.User) authentication.getPrincipal();
     UserData userData =
         userQueryService.findById(user.getId()).orElseThrow(ResourceNotFoundException::new);
-    UserWithToken userWithToken = new UserWithToken(userData, authorization.split(" ")[1]);
+    UserWithToken userWithToken = UserWithToken.of(userData, authorization.split(" ")[1]);
     User result =
         User.newBuilder()
-            .email(userWithToken.getEmail())
-            .username(userWithToken.getUsername())
-            .token(userWithToken.getToken())
+            .email(userWithToken.email())
+            .username(userWithToken.username())
+            .token(userWithToken.token())
             .build();
     return DataFetcherResult.<User>newResult().data(result).localContext(user).build();
   }
