@@ -2,10 +2,10 @@ package io.spring.infrastructure.mybatis.readservice;
 
 import io.spring.application.CursorPageParameter;
 import io.spring.application.data.CommentData;
+import java.time.Instant;
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-import org.joda.time.DateTime;
 
 @Mapper
 public interface CommentReadService {
@@ -14,5 +14,5 @@ public interface CommentReadService {
   List<CommentData> findByArticleId(@Param("articleId") String articleId);
 
   List<CommentData> findByArticleIdWithCursor(
-      @Param("articleId") String articleId, @Param("page") CursorPageParameter<DateTime> page);
+      @Param("articleId") String articleId, @Param("page") CursorPageParameter<Instant> page);
 }
